+++ packages/component-core/README.md
@@ -1,0 +1,18 @@
+# component-core
+
+XMPP component core for JavaScript
+
+## Install
+
+```
+npm install @xmpp/component-core
+```
+
+## Usage
+
+See [example.js](https://github.com/node-xmpp/node-xmpp/tree/master/packages/component/example.js)
+
+## References
+
+* [XEP-0114: Jabber Component Protocol](https://xmpp.org/extensions/xep-0114.html) implemented
+* [XEP-0225: Component Connections](https://xmpp.org/extensions/xep-0225.html) todo
