+++ packages/component-core/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@xmpp/component-core",
+  "description": "XMPP component core for JavaScript",
+  "repository": "github:node-xmpp/node-xmpp",
+  "homepage": "https://github.com/node-xmpp/node-xmpp/tree/master/packages/component-core",
+  "bugs": "http://github.com/node-xmpp/node-xmpp/issues",
+  "version": "0.0.0",
+  "license": "ISC",
+  "keywords": [
+    "XMPP",
+    "component",
+    "core"
+  ],
+  "dependencies": {
+    "@xmpp/connection-tcp": "^0.0.0",
+    "@xmpp/xml": "^0.0.1"
+  }
+}
