+++ packages/component/README.md
@@ -2,6 +2,8 @@
 
 XMPP component for JavaScript
 
+Includes `@xmpp/component-core` and `@xmpp/reconnect`.
+
 ## Install
 
 ```
@@ -11,8 +13,3 @@ npm install @xmpp/component
 ## Usage
 
 See [example.js](https://github.com/node-xmpp/node-xmpp/tree/master/packages/component/example.js)
-
-## References
-
-* [XEP-0114: Jabber Component Protocol](https://xmpp.org/extensions/xep-0114.html) implemented
-* [XEP-0225: Component Connections](https://xmpp.org/extensions/xep-0225.html) todo
